modules/location_event_consumer/requirements.txt
kafka-python == 2.0.2
psycopg2-binary==2.8.5
SQLAlchemy==1.3.19
requests==2.25.0